classes/meson: dump the end of the meson-log.txt on failure
When 'meson setup' fails it just prints the summary of the failure and
leaves the full details in the meson-log.txt.

  ../sources/xorgproto-2024.1/meson.build:22:0: ERROR:
  Executables created by c compiler x86-poky-linux-gcc [...] are not runnable.

  A full log can be found at [${B}]/meson-logs/meson-log.txt
  ERROR: meson failed

In CI systems where the build tree is then thrown away this is not very
useful, so on failure also output the last ten lines of the log file:
hopefully that is enough for context without flooding the logs.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>