base-passwd: add clock group
New systemd version (v258) introduces a new group, clock, to
"enable applications like linuxptp to open clocks without root
privileges".[1]

This results in warning at do_rootfs time:
WARNING: Group clock has never been defined

Add group clock with gid 81 to fix this issue. The wheel group's
gid is 80, so 81 is chosen.

Note that Debian rejects this patch because they're using sysusers
to handle such groups, including 'clock', 'kvm', 'sgx', etc. We're
not relying on sysusers, so we have to add a patch here.

Below are more detailed reasons for why these groups (wheel, sgx, kvm,
clock) need to be in base-passwd as oe-specific patches instead of being
added via separate recipes:
OE currently does not have a mechanism like systemd-sysusers that if a
group has been added, the related configurations in separate packages
do not have confliction.
In OE, when multiple recipes need the same group, and these recipes
do not necessarily have deps on each other, the current viable way
to do this is to add that group to base-passwd.
If things change in the future, which means significant change to our
useradd related bbclasses, we might do similar as Debian. But before
that, these groups need to remain in base-passwd.

[1] https://github.com/systemd/systemd/commit/af96ccfc24bc4803078a46b4ef2cdeb5decdfbcd

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>