lib/oe/patch: use author date as commit date for "git am"
By default, "git am" uses the time of commit creation as the committer
date and is therefore "unique" to each build.

Use the --committer-date-is-author-date option for git to use the author
date as the committer date to get deterministic meta-data and therefore
commit hashes (which may end-up in sources or build in the form of e.g.
localversion file).

Cc: isar-users@googlegroups.com
Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>