recipetool: support PEP639-variant of license key in pyproject.toml
Python modules using PEP639-variant of the license key[1] are currently
returning Unknown as LICENSE when using recipetool create on them
because we try to parse the key as a dict but it is now an SPDX license
expression.

This adds support for PEP639-variant of the license key, though it does
not handle the newly added license-files key[2] as I couldn't find a
dual-licensed recipe which uses that mechanism yet.

[1] https://peps.python.org/pep-0639/#add-string-value-to-license-key
[2] https://peps.python.org/pep-0639/#add-license-files-key

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>