core-image-ptest: drop PTEST_EXPECT_FAILURE
This variable setting is not helping people now. Because in practice,
ptest failure warnings on autobuilder are still catched and reported
back to developer.

In addition, setting this value to 1 is sometimes causing confusion for
developers.

As an example, I used 'bitbake core-image-ptest-util-linux:do_testimage'.
It succeeded with warning message. I didn't notice the warning message.
I saw the command succeeded and I tought util-linux ptest is OK. But
in actual fact, the ptest failed. It's the PTEST_EXPECT_FAILURE setting
in this core-image-ptest.bb that is not giving me error.

In summary, a simple erroring out action should be clearer.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>