classes/meson: clean up path arguments
The --bindir (etc.) arguments can be absolute paths, so there's no need
to work out the relative paths and we can instead pass them directly.

Also re-order the options to match the order in the Meson manual, for
ease of checking that we pass all the paths that Meson uses.

Notably, we don't set --licensedir, but we handle license manifests
ourselves.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>