run-postinsts: fix post-install-script output in log file
When the POSTINST_LOGGING is enabled the output of the post-install
script is not stored in the 'postinstall.log' file. This probably
happened since commit 706410c847ac9c89317d098de5d5c580736edbbb.

Fix the post-install script execution statement to honor the
stdout/stderr redirection to the 'postinstall.log' file.

Signed-off-by: Samuele Favazza <samuele.favazza@duagon.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>