kea: remove more host paths
If ccache is in use then the full path to ccache is resolved, and ends
up in both config.report and then config_report.cc.  This results in
build paths in libkea-cfgrpt.so.3.0.0, so we should also remove
HOSTTOOLS_DIR from these files to ensure both reproducibility and no
host contamination.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>