fragments/autobuilder: Add back PTEST_EXPECT_FAILURE
The PTEST_EXPECT_FAILURE option setting was recently dropped from
core-image-ptest [1]. This is probably the right thing to do, but we
still want to use it on the autobuilder.

Note, this will now be enabled on all images and not just the ptest
ones, but this should not change the whole behaviour.

[1] https://lists.openembedded.org/g/openembedded-core/message/227412

Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>