build-appliance-image: do not inherit setuptools
The recipe is using pip, and that doesn't require setuptools. I believe it
was a mix-up from https://bugzilla.yoctoproject.org/show_bug.cgi?id=11022
where the initial idea was to use easy_install from setuptools.

A particular reason to drop the setuptools inherit is that setuptools brings in
python3native class which sets PYTHONNOUSERSITE = "1" to avoid $HOME contamination.
This is causing errors with pip 25.3, which rejects --user option in presence
of that environment variable. This recipe redirects $HOME before running pip,
so contamination is avoided.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>